Fixed issue with n/ack after stopping persistence subscription.
Basically, without using a persistent subscription as a context manager, the stop() method needs to actually stop the gRPC call, meaning that a subsequent attempt to ack() a received event should raise a ProgrammingError. However, when a persistent subscription is used as a context manager, we can defer actually stopping the gRPC call to the exiting of the context manager and simply stop the persistent subscription iteration, meaning in this case that a received event can be n/acked.

(cherry picked from commit cf8144f64167f5152e2f03aa986edca51a90de30)